requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
Deprecated==1.3.1
deprecation==2.1.0
fastapi==0.124.0
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
mss==10.1.0
multidict==6.7.0
nodriver==0.48.1
packaging==25.0
playwright==1.57.0
postgrest==2.25.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
realtime==2.25.0
starlette==0.50.0
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
websockets==15.0.1
wrapt==2.0.1
yarl==1.22.0
gspread==6.0.0
oauth2client==4.1.3
